KubenAI-API/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic-settings
requests
